refactor: move canvasmanager stuff to its own directory
