+++ js/canvasmanager/drawobjects/drawobject.ts
