+++ tests/canvasmanager.test.ts
@@ -1,5 +1,5 @@
-import { CanvasManager } from "../js/canvasmanager";
-import { Circle } from "../js/drawobjects/circle";
+import { CanvasManager } from "../js/canvasmanager/canvasmanager";
+import { Circle } from "../js/canvasmanager/drawobjects/circle";
 
 describe("CanvasManager", () => {
   it("can undo clearing the canvas", () => {
